+++ sdk/ts-aws/src/index.ts
@@ -1,0 +1,11 @@
+export type {
+	KMSClient,
+	KMSGetPublicKeyInput,
+	KMSGetPublicKeyOutput,
+	KMSRequestOptions,
+	KMSSignerOptions,
+	KMSSignInput,
+	KMSSignOutput,
+	Signer,
+} from "./signer.js";
+export { KMSSigner } from "./signer.js";